Changing to es6 syntax and reorganization of files
