+++ push-notifications/app.js
@@ -1,21 +1,19 @@
-let express = require('express');
-let app = express();
-let pool = require('./database');
+import express from 'express';
+import kafka from 'kafka-node';
+import * as kafkaStreams from './src/handlers/kafkaEventStreamHandler';
+import indexRouter from './src/routes/index';
+import tokenRouter from './src/routes/tokens';
 
-let kafka = require('kafka-node');
-let kafkaStreams = require('./src/handlers/kafkaEventStreamHandler');
 const Consumer = kafka.Consumer;
 const client = new kafka.KafkaClient("localhost:2181");
+let app = express();
 
 //Handle push notifications
 
 const consumer = new Consumer(client, [ { topic: 'employee', partition: 0 } ], { autoCommit: true });
-consumer.on('message', function (message) {
+consumer.on('message', (message) => {
     kafkaStreams.handleKafkaEvents(message);
 });
-
-const indexRouter = require('./src/routes/index');
-const tokenRouter = require('./src/routes/tokens');
 
 const PORT_NUMBER = 3000; // port on which you want to run your server on
 
@@ -29,4 +27,4 @@ app.listen(PORT_NUMBER, () => {
     console.log('Server Online on Port' + PORT_NUMBER);
 });
 
-module.exports = app;
+export default app;
