+++ push-notifications/src/handlers/pushNotificationsHandler.js
@@ -1,16 +1,15 @@
-const { Expo } = require('expo-server-sdk');
-let pool = require('../../database');
-let db_config = require('../../db_config');
-
-const expo = new Expo();
+import {Expo} from 'expo-server-sdk';
+import pool from '../conf/database';
+import {db} from '../conf/config';
 
 // let savedPushTokens = ['ExponentPushToken[P6Hm9oEChVDz7OPwO-_Qsj]'];
 
+const expo = new Expo();
+const table = db.table;
 let savedPushTokens = [];
-const table = db_config.table;
 
-const handlePushTokens = (message) => {
-    pool.query(`SELECT * FROM ${table}`, function(err, response){
+export const handlePushTokens = (message) => {
+    pool.query(`SELECT * FROM ${table}`, (err, response) => {
         savedPushTokens = JSON.parse(JSON.stringify(response));
         savedPushTokens = savedPushTokens.map(element => {
             return element.token;
@@ -42,5 +41,3 @@ const handlePushTokens = (message) => {
         })(); 
     });
 }
-
-module.exports.handlePushTokens = handlePushTokens;
