+++ push-notifications/src/routes/index.js
@@ -1,8 +1,8 @@
-let express = require('express');
+import express from 'express';
 let router = express.Router();
 
 router.get('/', (req, res) => {
     res.send('Push Notification Server Running'); 
 });
 
-module.exports = router;
+export default router;
